fix: split code
