+++ .github/workflows/generator/app/CustomAction/PostCINotifications.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE NamedFieldPuns #-}
 
-module CustomAction.PostCINotifications (Status (..), Mode (..), currentPullRequestDiffMode, Params (..), step) where
+module CustomAction.PostCINotifications (Status (..), Mode (..), DiffHashPair (..), PullRequestInfo (..), currentPullRequestDiffMode, Params (..), step) where
 
 import Data.Aeson (ToJSON (toJSON))
 import Data.Map qualified as M
@@ -8,16 +8,25 @@ import Workflow.GitHub.Actions qualified as GHA
 
 data Status = SuccessStatus | FailureStatus String
 
-data Mode = DiffMode {diffHeadSHA :: String, diffBaseSHA :: String, diffPRNumber :: String, diffPRTitle :: String} | BranchMode
+data DiffHashPair = DiffHashPair {diffBase :: String, diffHead :: String}
+
+data PullRequestInfo = PullRequestInfo {prTitle :: String, prNumber :: String}
+
+data Mode = DiffMode DiffHashPair PullRequestInfo | BranchMode
 
 currentPullRequestDiffMode :: Mode
-currentPullRequestDiffMode =
-  DiffMode
-    { diffHeadSHA = GHA.mkExpression "github.event.pull_request.head.sha",
-      diffBaseSHA = GHA.mkExpression "github.event.pull_request.base.sha",
-      diffPRNumber = GHA.mkExpression "github.event.number",
-      diffPRTitle = GHA.mkExpression "github.event.pull_request.title"
-    }
+currentPullRequestDiffMode = DiffMode hashPair pr
+  where
+    hashPair =
+      DiffHashPair
+        { diffBase = GHA.mkExpression "github.event.pull_request.base.sha",
+          diffHead = GHA.mkExpression "github.event.pull_request.head.sha"
+        }
+    pr =
+      PullRequestInfo
+        { prTitle = GHA.mkExpression "github.event.pull_request.title",
+          prNumber = GHA.mkExpression "github.event.number"
+        }
 
 data Params = Params {status :: Status, beginTime :: String, reportName :: String, mode :: Mode}
 
@@ -29,12 +38,12 @@ step params = GHA.namedAs "Notify" $ GHA.actionStep "Pctg-x8/ci-notifications-po
             SuccessStatus -> [("status", toJSON "success")]
             FailureStatus jobName -> [("status", toJSON "failure"), ("failure-step", toJSON jobName)]
           modeFields = M.fromList $ case mode params of
-            DiffMode {diffHeadSHA, diffBaseSHA, diffPRNumber, diffPRTitle} ->
+            DiffMode hashPair pr ->
               [ ("mode", toJSON "diff"),
-                ("head_sha", toJSON diffHeadSHA),
-                ("base_sha", toJSON diffBaseSHA),
-                ("pr_number", toJSON diffPRNumber),
-                ("pr_title", toJSON diffPRTitle)
+                ("head_sha", toJSON $ diffHead hashPair),
+                ("base_sha", toJSON $ diffBase hashPair),
+                ("pr_number", toJSON $ prNumber pr),
+                ("pr_title", toJSON $ prTitle pr)
               ]
             BranchMode -> [("mode", toJSON "branch")]
           commonParams = M.fromList [("begintime", toJSON $ beginTime params), ("report_name", toJSON $ reportName params)]
