+++ .github/workflows/generator/app/Utils.hs
@@ -1,0 +1,9 @@
+module Utils (applyModifiers, runOnFailure) where
+
+import Workflow.GitHub.Actions qualified as GHA
+
+applyModifiers :: (Foldable f) => f (a -> a) -> a -> a
+applyModifiers = flip $ foldl (flip ($))
+
+runOnFailure :: (GHA.Conditional c) => c -> c
+runOnFailure = GHA.withCondition "failure()"
