+++ .github/workflows/generator/package.yaml
@@ -42,6 +42,8 @@ default-extensions:
   - ImportQualifiedPost
   - TupleSections
   - FlexibleInstances
+  - FlexibleContexts
+  - MonoLocalBinds
 
 library:
   source-dirs: src
